face_recognition_sparseRep/requirements.txt
numba
opencv-python
matplotlib
numpy
cvxpy
scikit-learn
flask
mxnet
pyyaml
requests
black